requirements_dev.txt
-r requirements.txt

mock==2.0.0
pytest-docker-compose==1.0.1
pytest==3.6.0
pytest-asyncio==0.8.0
pytest-aiohttp==0.3.0
pytest-cov==2.5.1
coverage==4.5.1

flake8==3.5.0
mypy==0.560
bandit==1.4.0
Sphinx==1.6.7
sphinx-rtd-theme==0.2.4
watchdog==0.8.3
